+++ src/games/slashchain/components/game.tsx
